Rename mapstore to defaultstore
